requirements.txt
numpy
torch
lightning
hydra-core
mlflow
tensorboard
ipykernel
gdown